# docs/node_swagger.yaml
definitions:
...
  api.Block:
    properties:
      Header:
        $ref: '#/definitions/block.BlockHeader'
      ProducerSignature:
        items:
          type: integer
        type: array
      PubKeysBitmap:
        items:
          type: integer
        type: array
      Signature:
        items:
          type: integer
        type: array
      TxHashes:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      hash:
        type: string
      status:
        type: string
      transactions:
        items:
          $ref: '#/definitions/api.Transaction'
        type: array
    type: object
  api.Events:
    properties:
      address:
        type: string
      data:
        items:
          type: string
        type: array
      identifier:
        type: string
      topics:
        items:
          type: string
        type: array
    type: object
  api.Logs:
    properties:
      address:
        type: string
      contractId:
        type: integer
      events:
        items:
          $ref: '#/definitions/api.Events'
        type: array
    type: object
  api.Marketplace:
    properties:
      id:
        type: string
      name:
        type: string
      ownerAddress:
        type: string
      referralAddress:
        type: string
      referralPercentage:
        type: integer
    type: object
  api.Transaction:
    properties:
      Block:
        type: integer
      GasLimit:
        type: integer
      GasMultiplier:
        type: integer
      RawData:
        $ref: '#/definitions/transaction.Transaction_Raw'
      Receipts:
        items:
          $ref: '#/definitions/transaction.Transaction_Receipt'
        type: array
      Result:
        $ref: '#/definitions/transaction.Transaction_TXResult'
      ResultCode:
        $ref: '#/definitions/transaction.Transaction_TXResultCode'
      Signature:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      hash:
        type: string
      logs:
        $ref: '#/definitions/api.Logs'
      signature:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  asset.NFTAsset:
    properties:
      AdminAddress:
        items:
          type: integer
        type: array
      AssetType:
        $ref: '#/definitions/kapps.KDAData_EnumAssetType'
      Attributes:
        $ref: '#/definitions/kapps.AttributesData'
      BurnedValue:
        type: integer
      CirculatingSupply:
        type: integer
      ID:
        items:
          type: integer
        type: array
      InitialSupply:
        type: integer
      IssueDate:
        type: integer
      Logo:
        type: string
      MaxSupply:
        type: integer
      MintedValue:
        type: integer
      Name:
        items:
          type: integer
        type: array
      OwnerAddress:
        items:
          type: integer
        type: array
      Precision:
        type: integer
      Properties:
        $ref: '#/definitions/kapps.PropertiesData'
      Roles:
        items:
          $ref: '#/definitions/kapps.RolesData'
        type: array
      Royalties:
        $ref: '#/definitions/kapps.RoyaltiesData'
      Ticker:
        items:
          type: integer
        type: array
      URIs:
        additionalProperties:
          type: string
        type: object
      userKDA:
        $ref: '#/definitions/kapps.UserKDA'
    type: object
...
  block.BlockHeader:
    properties:
      BlockRewards:
        type: integer
      BurnedUnclaimed:
        type: integer
      ChainID:
        items:
          type: integer
        type: array
      Epoch:
        type: integer
      IsEpochStart:
        type: boolean
      KAppFees:
        type: integer
      KAppsTrieRoot:
        items:
          type: integer
        type: array
      Nonce:
        type: integer
      ParentHash:
        items:
          type: integer
        type: array
      PrevEpochStartSlot:
        type: integer
      PrevRandSeed:
        items:
          type: integer
        type: array
      RandSeed:
        items:
          type: integer
        type: array
      Reserved:
        items:
          type: integer
        type: array
      Slot:
        type: integer
      SoftwareVersion:
        items:
          type: integer
        type: array
      StakingRewards:
        type: integer
      Timestamp:
        type: integer
      TrieRoot:
        items:
          type: integer
        type: array
      TxBurnedFees:
        type: integer
      TxCount:
        type: integer
      TxFees:
        type: integer
      TxRootHash:
        items:
          type: integer
        type: array
      ValidatorsTrieRoot:
        items:
          type: integer
        type: array
    type: object
...
  kapps.AttributesData:
    properties:
      IsNFTMetadataChangeStopped:
        type: boolean
      IsNFTMintStopped:
        type: boolean
      IsPaused:
        type: boolean
      IsRoyaltiesChangeStopped:
        type: boolean
    type: object
  kapps.KDAData:
    properties:
      AdminAddress:
        items:
          type: integer
        type: array
      AssetType:
        $ref: '#/definitions/kapps.KDAData_EnumAssetType'
      Attributes:
        $ref: '#/definitions/kapps.AttributesData'
      BurnedValue:
        type: integer
      CirculatingSupply:
        type: integer
      ID:
        items:
          type: integer
        type: array
      InitialSupply:
        type: integer
      IssueDate:
        type: integer
      Logo:
        type: string
      MaxSupply:
        type: integer
      MintedValue:
        type: integer
      Name:
        items:
          type: integer
        type: array
      OwnerAddress:
        items:
          type: integer
        type: array
      Precision:
        type: integer
      Properties:
        $ref: '#/definitions/kapps.PropertiesData'
      Roles:
        items:
          $ref: '#/definitions/kapps.RolesData'
        type: array
      Royalties:
        $ref: '#/definitions/kapps.RoyaltiesData'
      Ticker:
        items:
          type: integer
        type: array
      URIs:
        additionalProperties:
          type: string
        type: object
    type: object
  kapps.KDAData_EnumAssetType:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - KDAData_Fungible
    - KDAData_NonFungible
    - KDAData_SemiFungible
  kapps.LastClaim:
    properties:
      Epoch:
        type: integer
      Timestamp:
        type: integer
    type: object
  kapps.PropertiesData:
    properties:
      CanAddRoles:
        type: boolean
      CanBurn:
        type: boolean
      CanChangeOwner:
        type: boolean
      CanFreeze:
        type: boolean
      CanMint:
        type: boolean
      CanPause:
        type: boolean
      CanWipe:
        type: boolean
      LimitTransfer:
        type: boolean
    type: object
  kapps.RolesData:
    properties:
      Address:
        items:
          type: integer
        type: array
      HasRoleDeposit:
        type: boolean
      HasRoleMint:
        type: boolean
      HasRoleSetITOPrices:
        type: boolean
      HasRoleTransfer:
        type: boolean
    type: object
  kapps.RoyaltiesData:
    properties:
      Address:
        items:
          type: integer
        type: array
      ITOFixed:
        type: integer
      ITOPercentage:
        type: integer
      MarketFixed:
        type: integer
      MarketPercentage:
        type: integer
      SplitRoyalties:
        additionalProperties:
          $ref: '#/definitions/kapps.RoyaltySplitData'
        type: object
      TransferFixed:
        type: integer
      TransferPercentage:
        items:
          $ref: '#/definitions/kapps.RoyaltyData'
        type: array
    type: object
  kapps.RoyaltyData:
    properties:
      Amount:
        type: integer
      Percentage:
        type: integer
    type: object
  kapps.RoyaltySplitData:
    properties:
      PercentITOFixed:
        type: integer
      PercentITOPercentage:
        type: integer
      PercentMarketFixed:
        type: integer
      PercentMarketPercentage:
        type: integer
      PercentTransferFixed:
        type: integer
      PercentTransferPercentage:
        type: integer
    type: object
  kapps.UserBucket:
    properties:
      Delegation:
        items:
          type: integer
        type: array
      StakedAt:
        type: integer
      StakedEpoch:
        type: integer
      UnstakedEpoch:
        type: integer
      Value:
        type: integer
    type: object
  kapps.UserKDA:
    properties:
      Balance:
        type: integer
      Buckets:
        additionalProperties:
          $ref: '#/definitions/kapps.UserBucket'
        type: object
      FrozenBalance:
        type: integer
      LastClaim:
        $ref: '#/definitions/kapps.LastClaim'
      MIME:
        items:
          type: integer
        type: array
      Metadata:
        items:
          type: integer
        type: array
    type: object
  kdafeespool.KDAFeesPoolData:
    properties:
      Active:
        type: boolean
      AdminAddress:
        items:
          type: integer
        type: array
      ConvertedFees:
        type: integer
      FRatioKDA:
        type: integer
      FRatioKLV:
        type: integer
      KDA:
        items:
          type: integer
        type: array
      KDABalance:
        type: integer
      KLVBalance:
        type: integer
      OwnerAddress:
        items:
          type: integer
        type: array
    type: object
...
  shared.GenericAPIResponse:
    properties:
      code:
        $ref: '#/definitions/shared.ReturnCode'
      data: {}
      error:
        type: string
    type: object
  shared.ReturnCode:
    enum:
    - successful
    - internal_issue
    - bad_request
    - system_busy
    type: string
    x-enum-varnames:
    - ReturnCodeSuccess
    - ReturnCodeInternalError
    - ReturnCodeRequestError
    - ReturnCodeSystemBusy
...
  transaction.TXContract:
    properties:
      Parameter:
        $ref: '#/definitions/anypb.Any'
      Type:
        $ref: '#/definitions/transaction.TXContract_ContractType'
    type: object
  transaction.TXContract_ContractType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    - 19
    - 20
    - 21
    - 22
    - 23
    - 24
    - 63
    format: int32
    type: integer
    x-enum-varnames:
    - TXContract_TransferContractType
    - TXContract_CreateAssetContractType
    - TXContract_CreateValidatorContractType
    - TXContract_ValidatorConfigContractType
    - TXContract_FreezeContractType
    - TXContract_UnfreezeContractType
    - TXContract_DelegateContractType
    - TXContract_UndelegateContractType
    - TXContract_WithdrawContractType
    - TXContract_ClaimContractType
    - TXContract_UnjailContractType
    - TXContract_AssetTriggerContractType
    - TXContract_SetAccountNameContractType
    - TXContract_ProposalContractType
    - TXContract_VoteContractType
    - TXContract_ConfigITOContractType
    - TXContract_SetITOPricesContractType
    - TXContract_BuyContractType
    - TXContract_SellContractType
    - TXContract_CancelMarketOrderContractType
    - TXContract_CreateMarketplaceContractType
    - TXContract_ConfigMarketplaceContractType
    - TXContract_UpdateAccountPermissionContractType
    - TXContract_DepositContractType
    - TXContract_ITOTriggerContractType
    - TXContract_SmartContractType
...
  transaction.Transaction_KDAFee:
    properties:
      Amount:
        description: 'TODO: allow spread'
        type: integer
      KDA:
        items:
          type: integer
        type: array
    type: object
  transaction.Transaction_Raw:
    properties:
      BandwidthFee:
        type: integer
      ChainID:
        items:
          type: integer
        type: array
      Contract:
        items:
          $ref: '#/definitions/transaction.TXContract'
        type: array
      Data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      KAppFee:
        type: integer
      KDAFee:
        $ref: '#/definitions/transaction.Transaction_KDAFee'
      Nonce:
        type: integer
      PermissionID:
        type: integer
      Sender:
        items:
          type: integer
        type: array
      Version:
        type: integer
    type: object
  transaction.Transaction_Receipt:
    properties:
      Data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
    type: object
  transaction.Transaction_TXResult:
    enum:
    - 0
    - 1
    format: int32
    type: integer
    x-enum-varnames:
    - Transaction_SUCCESS
    - Transaction_FAILED
...
info:
  contact: {}
  description: This is a sample of the KleverChain Node Server.
  title: KleverChain Node
  version: "1.0"
paths:
...
  /asset/{id}:
    get:
      parameters:
      - description: asset id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    asset:
                      $ref: '#/definitions/kapps.KDAData'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns an assetResponse containing information about the asset
      tags:
      - Asset
  /asset/nft/{owner}/{id}:
    get:
      parameters:
      - description: asset owner
        in: path
        name: owner
        required: true
        type: string
      - description: asset id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    asset:
                      $ref: '#/definitions/asset.NFTAsset'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns an assetResponse containing information about the asset
      tags:
      - Asset
  /asset/pool/{id}:
    get:
      parameters:
      - description: asset id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    pool:
                      $ref: '#/definitions/kdafeespool.KDAFeesPoolData'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns an poolResponse containing information about the asset pool
      tags:
      - Asset
  /block/by-hash/{hash}:
    get:
      parameters:
      - description: block hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    block:
                      $ref: '#/definitions/api.Block'
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the block info by hash
      tags:
      - Block
  /block/by-nonce/{nonce}:
    get:
      parameters:
      - description: block nonce
        in: path
        name: nonce
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    block:
                      $ref: '#/definitions/api.Block'
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the block info by nonce
      tags:
      - Block
  /marketplace/{id}:
    get:
      parameters:
      - description: marketplace Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    marketplace:
                      $ref: '#/definitions/api.Marketplace'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns an marketplaceResponse containing information about the marketplace
      tags:
      - Marketplace
...
  /network/network-parameters:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    parameters:
                      type: object
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the active parameters of the proposal controller
      tags:
      - Network
...
swagger: "2.0"